+++ Source/Engine/src/Utilities/DeltaTime.h
@@ -9,12 +9,12 @@ namespace PAL
 		
 		void Update();
 
-		inline float GetDeltaInMilliSeconds() const { return m_DeltaTime.FrameTime * 1000.0; }
+		inline float GetDeltaInMilliSeconds() const { return m_DeltaTime.FrameTime * 1000.0f; }
 		inline float GetDeltaInSeconds() const { return m_DeltaTime.FrameTime; }
 
 		inline int GetAverageFPS() const { return m_AvgData.FPS; }
 		inline float GetAvgDeltaTimeInSeconds() const { return m_AvgData.AvgFrameTime; }
-		inline float GetAvgDeltaTimeInMilliSeconds() const { return m_AvgData.AvgFrameTime * 1000.0; }
+		inline float GetAvgDeltaTimeInMilliSeconds() const { return m_AvgData.AvgFrameTime * 1000.0f; }
 	private:
 		
 		struct DeltaTimeData
@@ -27,7 +27,7 @@ namespace PAL
 		{
 			float LastFrameTime = 0.0f;
 			float AvgFrameTime = 0.0f;
-			double TotalFrameTimeInOneSec = 0.0f;
+			float TotalFrameTimeInOneSec = 0.0f;
 
 			int FPS = 0;
